Add files via upload
We will use decision tree from scikit learn library to generate a random forest and use it for loan approval classification(to approve or not to) to an individual. Obviously there is random forest built in function available for this. But then whats fun in that??
Libraries needed: 
1. Scikit Learn
2. Random
3. CSV reader 

Steps:
1. Import libraries
2. Load data set from CSV
3. Clean data set
4. Form bags (small data sets for each tree)
5. Initialize trees and pass them the bags for training
6. Pass a new data-point for prediction.

System Requirements:
Python 3.0 with above mentioned libraries installed.
The "German Credit Data-set" as csv file (included in this folder)